requirements.txt
asgiref==3.3.4
class-registry==2.1.2
dj-database-url==0.5.0
Django==3.2.3
django-filter==2.4.0
docopt==0.6.2
filters==1.3.2
gunicorn==20.1.0
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
qs==1.4.1
regex==2021.4.4
six==1.16.0
sqlparse==0.4.1
whitenoise==5.2.0
psycopg2==2.8.6
django-social-share==2.2.1
django-heroku==0.3.1

